requirements.txt
scipy
numpy
torch
pathlib
tensorflow
g2p-en
pytorch-ignite
fire